requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
python-dotenv
PyYAML
flask
-e .